test: various test reliability improvements
Make fit save/load file dialogs behave automatically under pytest

Refactor intereactive test module structure

Convert uses of `qtbot.wait_signal` to correct syntax using context manager

Increase timeout for wait_until calls in explorer test
