+++ tests/interactive/imagetool/test_imagetool.py
@@ -489,8 +489,8 @@ def test_itool_rotate(qtbot, accept_dialog) -> None:
         assert dialog.angle_spin.value() == 30.0
         assert dialog.center_spins[0].value() == 3.0
         assert dialog.center_spins[1].value() == 3.1
-        dialog.copy_button.click()
-        qtbot.wait_signal(dialog._sigCodeCopied)
+        with qtbot.wait_signal(dialog._sigCodeCopied):
+            dialog.copy_button.click()
         dialog.reshape_check.setChecked(True)
         dialog.new_window_check.setChecked(False)
 
@@ -536,8 +536,8 @@ def test_itool_crop_view(qtbot, accept_dialog) -> None:
     def _set_dialog_params(dialog: CropToViewDialog) -> None:
         dialog.dim_checks["x"].setChecked(True)
         dialog.dim_checks["y"].setChecked(True)
-        dialog.copy_button.click()
-        qtbot.wait_signal(dialog._sigCodeCopied)
+        with qtbot.wait_signal(dialog._sigCodeCopied):
+            dialog.copy_button.click()
         dialog.new_window_check.setChecked(False)
 
     _handler = accept_dialog(win.mnb._crop_to_view, pre_call=_set_dialog_params)
@@ -590,8 +590,8 @@ def test_itool_crop(qtbot, accept_dialog) -> None:
         dialog.cursor_combos[1].setCurrentIndex(2)
         dialog.dim_checks["x"].setChecked(True)
         dialog.dim_checks["y"].setChecked(True)
-        dialog.copy_button.click()
-        qtbot.wait_signal(dialog._sigCodeCopied)
+        with qtbot.wait_signal(dialog._sigCodeCopied):
+            dialog.copy_button.click()
         dialog.new_window_check.setChecked(False)
 
     _h0 = accept_dialog(win.mnb._crop, pre_call=_set_dialog_params)
@@ -611,8 +611,8 @@ def test_itool_crop(qtbot, accept_dialog) -> None:
         dialog.cursor_combos[1].setCurrentIndex(2)
         dialog.dim_checks["x"].setChecked(True)
         dialog.dim_checks["y"].setChecked(False)
-        dialog.copy_button.click()
-        qtbot.wait_signal(dialog._sigCodeCopied)
+        with qtbot.wait_signal(dialog._sigCodeCopied):
+            dialog.copy_button.click()
         dialog.new_window_check.setChecked(False)
 
     _h1 = accept_dialog(win.mnb._crop, pre_call=_set_dialog_params)
@@ -641,8 +641,8 @@ def test_itool_average(qtbot, accept_dialog) -> None:
     # Test dialog
     def _set_dialog_params(dialog: AverageDialog) -> None:
         dialog.dim_checks["x"].setChecked(True)
-        dialog.copy_button.click()
-        qtbot.wait_signal(dialog._sigCodeCopied)
+        with qtbot.wait_signal(dialog._sigCodeCopied):
+            dialog.copy_button.click()
         dialog.new_window_check.setChecked(False)
 
     _handler = accept_dialog(win.mnb._average, pre_call=_set_dialog_params)
@@ -670,8 +670,8 @@ def test_itool_symmetrize(qtbot, accept_dialog) -> None:
     def _set_dialog_params(dialog: SymmetrizeDialog) -> None:
         dialog._dim_combo.setCurrentIndex(2)
         dialog._center_spin.setValue(2.0)
-        dialog.copy_button.click()
-        qtbot.wait_signal(dialog._sigCodeCopied)
+        with qtbot.wait_signal(dialog._sigCodeCopied):
+            dialog.copy_button.click()
         dialog.new_window_check.setChecked(False)
 
     _handler = accept_dialog(win.mnb._symmetrize, pre_call=_set_dialog_params)
