+++ tests/interactive/imagetool/test_imagetool_coordinate_widget.py
