+++ tests/interactive/test_derivative.py
@@ -1,0 +1,50 @@
+import numpy as np
+import pytest
+import xarray as xr
+
+import erlab
+from erlab.interactive.derivative import DerivativeTool, dtool
+
+
+@pytest.mark.parametrize("method_idx", [0, 1, 2, 3, 4])
+@pytest.mark.parametrize("interpmode", ["interp", "nointerp"])
+@pytest.mark.parametrize(
+    ("smoothmode", "nsmooth"),
+    [
+        ("none", 1),
+        ("gaussian", 1),
+        ("gaussian", 3),
+        ("boxcar", 1),
+        ("boxcar", 3),
+    ],
+)
+def test_dtool(qtbot, interpmode, smoothmode, nsmooth, method_idx) -> None:
+    data = xr.DataArray(np.arange(25).reshape((5, 5)), dims=["x", "y"]).astype(
+        np.float64
+    )
+    win: DerivativeTool = dtool(data, execute=False)
+    qtbot.addWidget(win)
+
+    def check_generated_code(w: DerivativeTool) -> None:
+        namespace = {"era": erlab.analysis, "data": data, "np": np, "result": None}
+        exec(w.copy_code(), {"__builtins__": {"range": range}}, namespace)  # noqa: S102
+        xr.testing.assert_identical(w.result, namespace["result"])
+
+    win.interp_group.setChecked(interpmode == "interp")
+    win.smooth_group.setChecked(smoothmode != "none")
+    win.sn_spin.setValue(nsmooth)
+
+    match smoothmode:
+        case "gaussian":
+            win.smooth_combo.setCurrentIndex(0)
+        case "boxcar":
+            win.smooth_combo.setCurrentIndex(1)
+
+    win.tab_widget.setCurrentIndex(method_idx)
+    if method_idx == 1:
+        win.lapl_factor_spin.setValue(40)
+    elif method_idx == 3:
+        win.curv_factor_spin.setValue(40)
+
+    check_generated_code(win)
+    win.close()
